+++ .devops/repo/src/commands/test.rs
@@ -1,4 +1,4 @@
-use crate::utils::{read_applications, repo_root};
+use crate::utils::{execute, print_application_errors, read_applications, repo_root};
 use std::process::Command;
 
 pub(crate) fn run() {
@@ -13,27 +13,17 @@ pub(crate) fn run() {
         // TODO: safety
         let project = std::fs::canonicalize(format!("{}/{}/project", root, app)).unwrap();
 
-        let result = Command::new("cargo")
-            .current_dir(project)
-            .arg("test")
-            .arg("--color")
-            .arg("always")
-            .arg("-q")
-            .status();
-        match result {
-            Ok(status) => {
-                if !status.success() {
-                    errors.push(app.clone());
-                }
-            }
-            Err(_) => errors.push(app.clone()),
-        }
+        execute(
+            Command::new("cargo")
+                .current_dir(project)
+                .arg("test")
+                .arg("--color")
+                .arg("always")
+                .arg("-q"),
+            &mut errors,
+            &app,
+        );
     }
 
-    if 0 < errors.len() {
-        println!("\nErrors found in");
-        for app in errors.iter() {
-            println!("    {}", app);
-        }
-    }
+    print_application_errors(errors, "Errors found in".to_string());
 }
